Move image_dir fixture to a shared location.
